path normalization to commons.
